requirements.txt
fastapi==0.115.6
langchain==0.3.12
pylint==3.3.2
langchain-community==0.3.12
python-dotenv==1.0.1
faiss-cpu>=1.9.0.post1
jinja2==3.1.4
python-multipart==0.0.19
uvicorn>=0.33.0
pdfplumber>=0.11.4
torch==2.2.2
sentence-transformers>=3.3.1
langchain-huggingface>=0.1.2
numpy<2.0
sentencepiece>=0.2.0
accelerate>=0.26.0
datasets>=3.2.0
pandas==2.2.3